Adding an endpoint to list the rooms
Adding an endpoint in the pusher to list all the rooms.
The pusher will contact ALL back servers and get the list of rooms from there.
It aggregates the rooms and sends it back to the caller.

The back server only has 1 second to answer, otherwise, the answer is ignored.

The end-point is protected via the Admin token.